// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/door-link-ayah-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { LinkingResolvedWordDto } from '../models/linking-resolved-word-dto';
export interface DoorLinkAyahDto {
  ayahId: number;
  ayahNumber: number;
  descriptions: Array<string>;
  pageFrom: number;
  pageTo: number;
  selectedWordIds: Array<number>;
  surahNameArabic: string;
  surahNumber: number;
  verseKey: string;
  words: Array<LinkingResolvedWordDto>;
}